Add file approval and fine-tuned permissions; bump to v1.0
This commit drastically changes the functionality of Approved Revs.
It is now possible to define specifically who can approve certain
namespaces, categories and pages. Additionally, it is possible to
define the approver as the person who created the page, the user
who "owns" the page (in the User namespace), and if Semantic
MediaWiki is installed, by properties set in pages.

Additionally, it is now possible to approve files. This affects
which version of the file is shown in the page.

ApprovedRevs.hooks.php:
* Added several hooks to handle file approvals
* Fixed cache purge issue which wasn't clearing history page cache on approval
  * This causes a bug in SMW 1.9.0 only. Not in 1.9.1+
* Made ApprovedRevs_Files.sql get called in update.php

ApprovedRevs_body.php:
* significant changes to entire file, including:
  * modified pageIsApprovable() to use new $egApprovedRevsPermissions
  * added mediaIsApprovable() to handle files
  * modified userCanApproved() to use new $egApprovedRevsPermissions
  * Added several methods

specials/
* Moved SpecialApprovedRevs.php to specials/SpecialApprovedPages.php
* QueryPage broken out into specials/SpecialApprovedPagesQueryPage.php
* Created specials/SpecialApprovedFiles.php:
  * Largely a clone of SpecialApprovedPages.php
* Created specials/SpecialApprovedFilesQueryPage.php
  * New file to handle Special:ApprovedFiles queries

README:
* bump version to 1.0

ApprovedRevs.alias.php:
* new aliases to account for new special page Special:ApprovedFiles

maintenance/
* ApprovedRevs_Files.sql: New file for file approvals SQL
* ApprovedRevs.sql: moved file from /ApprovedRevs.sql

i18n/
* en.json: Added messages to support new features
* qqq.json: Added message descriptions

Change-Id: I81746a7a0671d23f938cc3c2479544eadc59768e